+++ src/control_room/data_io.hpp
@@ -7,7 +7,7 @@ class DataIO
 {
 public:
     virtual void write_data(const char *data) = 0;
-    virtual ssize_t read_data(char *&buffer, size_t size) = 0;
+    virtual int read_data(char *&buffer, size_t size) = 0;
 
     virtual ~DataIO() {}
 };
